migration

+++ src/main/java/com/migration/migration/PostService.java
@@ -3,7 +3,7 @@ package com.migration.migration;
 import java.util.List;
 
 public interface PostService {
-    PostDTO getPost(Long id);
+    PostDTO getPost(int id);
     List<PostDTO> getPosts();
     List<PostDTO> savePosts();
 }
